+++ changelog.md
@@ -2,4 +2,5 @@
 
 
 
+- print version number
 - raise exception on error in subprocess